requirements.txt
aiocache==0.11.1
aiofiles==0.8.0
aiohttp==3.7.4.post0
APScheduler==3.8.1
astroid==2.11.6
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
chardet==4.0.0
colorthief==0.2.1
dill==0.3.5.1
discord.py @ git+https://github.com/Rapptz/discord.py@06df0d7931385a5b8efc1dc6bea79cc389b22808
dnspython==2.2.1
e==1.4.5
emoji-country-flag==1.3.0
expiringdict==1.2.1
fold-to-ascii==1.0.2.post1
humanfriendly==10.0
humanize==3.12.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mongoengine==0.23.1
msgpack==1.0.2
multidict==5.2.0
Pillow==9.1.1
platformdirs==2.5.2
psutil==5.8.0
pycodestyle==2.8.0
pylint==2.14.3
pymongo==3.12.1
python-dotenv==0.19.1
pytimeparse==1.1.8
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
ujson==4.2.0
wrapt==1.14.1
yarl==1.7.0
